Created a syntax_test file, and began process of rewriting for arbitrary command highlighting
Currently the syntax only matches simple commands, variables, and strings. String matching is about to be rewritten and drastically simplified

It seems that allowing arbitrary commands to be matched is going to require refactoring the entire essence of the syntax. Much of it appears to have been brought in directly from the Bash highlighting that ships with ST, which is overkill since Bash had a lot of edge cases where fish aims for overall simplicity