Added support for index range expansion of variables
In doing so, found possible bug with fish (see fish-shell/fish-shell#3339). I'd rather not have to write syntax to identify those cases as very specifically being invalid

Support for index range expansion of command substitution is on the way. It must be implemented separately, since spotting that a variable or command substitution will need to be in the item-access scope are fairly unique cases and it's not possible to make a generic item-access pattern