Upgraded command-call to match commands named with arbitrary strings
The implementation is a little limited based on the constraints of the ST regex engine. Even though the command name of a command call could span multiple lines using well-placed strings and escapes, we cannot match this condition. A Sublime Syntax file could, though, so that may be a future upgrade