Overhaul (and overcomplicate) command-call rules
This started as a little bit of refactoring to make piping easier to deal with, but finished as a large-scale rethink of how command calls and operators are matched. The fact this isn't a sublime-syntax file also shows up prominently because the amount of regex duplication in the command call matches has risen steeply once more. These could be rendered *so much more simply* if I had a scope stack