Dramatically restructure command-call again, introducing the concept of a command-chain
This allows me to finally treat pipes the way I want to, as a control operator which terminates a command call but does not terminate the overall command chain. This has already made piping simpler

The overall aim here is to making piping around blocks easy. I've currently rewritten the `begin`-`end` block to demonstrate how it could be done and the result is promising