Fix bug where '=' could not safely be a command name
I also switched to a much safer internal approach to needing '='
unmatched by the #string-unquoted rule such that the long option
argument pattern can find it. Instead of excluding '=' from all
unquoted strings in general and requiring manual treatment of it
every time, just recreate a modified #string-unquoted rule in the
one instance it is required!