Improve consistency of numeric scoping and recognise a + or - sign
Previously, any unquoted string that began with a number could be
scoped as constant.numeric, which made calling a function that was
just a number look a bit odd. Hence, now numeric matching is a part
of argument rather than string

Index expansion allows integers to have a sign, so the sign should be
recognised within this scope. It thus follows quite simply that the
sign be recognised in the command call scope as well

Prior to the end of options, a dash followed by an integer will be
considered an option rather than a numeric. Unfortunately it is up to
the implementation of individual programs whether such an argument
should be considered an option or a numeric, and we cannot decide
that. After the end of options "--" a dash followed by an integer is
considered a numeric, which should have a desirable effect in any
programs which consider this distinction important