Make significant improvements to handling of braces, change some names
I've renamed the scopes relating to brace expansion (or its lack thereof)
which has made them a lot clearer to understand. I've also renamed the
corresponding issues in the compatibility highlighter, and updated
the highlighter tests. I added some more syntax and highlighter tests
to capture new cases I found while doing this work

I also determined that a closing brace '}' is a completely invalid character
in command names for function calls, so I have added a match to handle it

Plus, when fixing up the brace expansion, I got to really flex the power
of the .sublime-syntax format. It's a good start to using that more
throughout the syntax