Rename double-ampersand and -bar scopes to distinguish in the meta part
Instead of being
  meta.function-call.operator.control keyword.operator.control.FOOBAR
now they are
  meta.function-call.operator.control.FOOBAR keyword.operator.control

This is for consistency with what I just did to the ampersand and semicolon,
which in turn is mostly for consistency with something I did to the redirection
operators. It's hard to decide which of these patterns is correct so for a long
time I have avoided even trying to make the choice. Now I'm fully committed to
one. Hope it was correct.