fish 3.4: Mark as deprecated an '&' without a separator before or after
The only practical means to achieve this is with the heretofore-unused regex
matching capability of the CompatHighlighter. While it is fairly trivial to
match ampersands that are not followed by a separator with just the
sublime-syntax alone, it is a much more complicated process to also ensure they
don't have a separator in front of them either. That is, without using a
lookbehind, which would cause the entire syntax schema to use the less
efficient parser. So, doing some external regex tests on all ampersands is the
trade-off that we make.